searchapi/requirements.txt
Django==1.11.13
django-filter==1.1.0
djangorestframework==3.8.2
Pillow==5.1.0
pytz==2018.4